chore: rename firebaseAuth to authService
